requirements.txt
APScheduler==3.10.4
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
PlexAPI==4.15.16
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
urllib3==2.2.3
Werkzeug==3.0.4
zipp==3.20.2
